+++ src/Propietario.cpp
@@ -5,7 +5,7 @@
 #include "Propietario.h"
 using namespace std;
 
-Propietario::Propietario(int id, string nombre, string sexo, string nacimiento, int puntaje, int disponibilidad): id(id), nombre(nombre), sexo(sexo), nacimiento(nacimiento), puntaje(puntaje), disponibilidad(disponibilidad){}
+Propietario::Propietario(int id, string nombre, string sexo, string nacimiento, float puntaje, int disponibilidad): id(id), nombre(nombre), sexo(sexo), nacimiento(nacimiento), puntaje(puntaje), disponibilidad(disponibilidad){}
 
 void Propietario::agregarHogar(string direccion, string descripcion, int camas, int bebes){
     this->pHogar = new Hogar(direccion, descripcion, camas, bebes);
@@ -39,7 +39,7 @@ string Propietario::getNacimiento(){
     return nacimiento;
 }
 
-void Propietario::setPuntaje(int puntaje){
+void Propietario::setPuntaje(float puntaje){
     this->puntaje = puntaje;
 }
